Split standards into two
