+++ acres/model/__init__.py
@@ -1,4 +1,4 @@
 """
 Package containing domain models (from the MVC design pattern).
 """
-__all__ = ['reference']
+__all__ = ['detection_standard', 'expansion_standard']
