+++ acres/model/detection_standard.py
